- upgraded code base to C++17.
Mainly to allow using 'if constexpr'.
Older CI compiler setups like GCC 5 for Linux were removed as a consequence.
Windows 32 bit was also removed because there are no plans to do any more 32 bit releases of GZDoom.

+++ src/common/engine/sc_man.cpp
@@ -1247,7 +1247,7 @@ void FScanner::AddSymbol(const char* name, uint64_t value)
 	Symbol sym;
 	sym.tokenType = TK_UIntConst;
 	sym.Number = value;
-	sym.Float = value;
+	sym.Float = (double)value;
 	symbols.Insert(name, sym);
 }
